# MM-29980: Optimize profilesInChannels cache to fast path