--- a/services/cache/lru.go
+++ b/services/cache/lru.go
@@ -221,6 +221,11 @@ func (l *LRU) get(key string, value interface{}) error {
 			_, err := s.UnmarshalMsg(e.value)
 			*v = &s
 			return err
+		case *map[string]*model.User:
+			var u model.UserMap
+			_, err := u.UnmarshalMsg(e.value)
+			*v = u
+			return err
 		}
 
 		// Slow path for other structs.