--- a/services/cache/lru_test.go
+++ b/services/cache/lru_test.go
@@ -279,6 +279,17 @@ func TestLRUMarshalUnMarshal(t *testing.T) {
 	// This does not make an actual difference in terms of functionality.
 	u.Timezone = nil
 	require.Equal(t, user, u)
+
+	tt := make(map[string]*model.User)
+	tt["1"] = u
+	err = l.Set("mm", model.UserMap(tt))
+	require.Nil(t, err)
+
+	var out map[string]*model.User
+	err = l.Get("mm", &out)
+	require.Nil(t, err)
+	out["1"].Timezone = nil
+	require.Equal(t, tt, out)
 }
 
 func BenchmarkLRU(b *testing.B) {
@@ -429,6 +440,39 @@ func BenchmarkLRU(b *testing.B) {
 			require.Nil(b, err)
 
 			var val model.User
+			err = l2.Get("test", &val)
+			require.Nil(b, err)
+		}
+	})
+
+	uMap := map[string]*model.User{
+		"id1": &model.User{
+			Id:       "id1",
+			CreateAt: 1111,
+			UpdateAt: 1112,
+			Username: "user1",
+			Password: "pass",
+		},
+		"id2": &model.User{
+			Id:       "id2",
+			CreateAt: 1113,
+			UpdateAt: 1114,
+			Username: "user2",
+			Password: "pass2",
+		},
+	}
+
+	b.Run("UserMap=new", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			l2 := NewLRU(&LRUOptions{
+				Size:                   1,
+				DefaultExpiry:          0,
+				InvalidateClusterEvent: "",
+			})
+			err := l2.Set("test", model.UserMap(uMap))
+			require.Nil(b, err)
+
+			var val map[string]*model.User
 			err = l2.Get("test", &val)
 			require.Nil(b, err)
 		}