@streamer45 2020-10-29
@agnivade FYI, there's a lint (gofmt) complain on `lru_test.go`.